refactor: streamline prod and test webapps route declaration
@mebster

Previously test webapp was directly using the underlying page generation
functions for each of the pages. Now it uses the prod entry points: the
functions '-toHtml and '-toHtmlError from the namespace
'slipstream.ui.views.representation.

This removes duplication and ensures that the code being tested with the
local test webapp is the same as in production.

Moreover, all pages have been refactored to share the same code
structure:

- view in namespace 'views.X' with a (page [metadata type]) entry point
- model in namespace 'models.X' with a (parse [metadata]) entry point
- test data in namespace 'models.X-test' with a 'raw-metadata var used
  in the test webapp
- expectations of the parsed 'raw-metadata in the same namespace
  'models.X-test'

This coherence in the code structure allows a much more straight forward
declaration of the correspondence between the 'pagenames' requested by
the server and the page generation functions for each view. At the same
time, this simplifies as well the declaration of the routes in the test
webapp removing a lot of code duplication.