fix: configuration parameters were parsed twice
@schaubl

The server always includes the configuration parameters into the
metadata before passing it to the UI module. Therefore, for the
configuration page, the parameters are duplicated.

The original mockup metadata correctly included these replication of the
parameters, but I removed it after seeing the duplication.

I've now included back the correct metadata with the parameter
duplication and fixed how the configuration parameters are parsed.