refactor: views can provide CSS and JS filenames without HTML template
@schaubl FYI

Previously, 'dashboard' and 'run' pages required an HTML template just
to provide the needed CSS and JS files. The HTML template did contain
anything else. See files deleted in this commit.

This change will easily allow views to load specific CSS and JS files.

Now the CSS and JS files can be also provided as strings in the
'context' map provided as argument to the function
slipstream.ui.views.base/generate.

This means that now, CSS and JS files are referenced in the
generated HTML file in the strict following order:

  1. Files referenced in the base.html template (which are always
  present).

  2. Files referenced in the HTML template used for a give view.

  3. Files provided in the 'html-dependencies' in the 'context' map
  mentioned above.

Note that in the HTML template files, the CSS files are referenced in the
HTML head, and the JS files in div#bottom-scripts at the bottom of the
file to allow pages to be displayed faster.