new: retrieve AJAX error detail from HTML error page in the response
Currently we cannot request the server with a JSON response to the AJAX
form submissions because when the request is successfull, the server
always sends the redirected HTML page, and jQuery fails parsing it as a
JSON.

Now, if the JSON response is not available, we try to find the error
detail in the header subtitle of the complete HTML error page in the
response.

Note that if we change to JSON, the JSON response will be automatically
parsed.