fix: correctly display the fields for new and old passwords.
@schaubl

Fixes "User creation with password: Error 409 - Conflict: Password
cannot be changed to an empty value."

Please note that the "old password" field is *only* required to modify
your own password. The only users that can modify other user profiles
are supers, and they do not need to provide the old password of the
profile when updating it. However supers still need to provide the old
password when updating their own password.