change: do not include HTML for secondary actions if not needed
Previously, following two actions where done on the JS side, after page
loading:

  - Remove secondary menu if no actions
  - Remove secondary menu dropdown caret if no extra actions

This caused some flicker of the dropdown caret since it was seen
sometimes before the JS had the chance to remove it.

Now the JS script code to handle that is removed because the HTML is not
generated by the clojure side is no need to.