fix: reports refresh in run page
@mebster @schaubl

Note that the refresh is done together with the refresh of the runtime
parameters, which happens every 15 secs only when the browser is in
use (i.e. has focus) and it stops when the run enters a final state.

Further, each time the browser is activated (i.e. gains focus) the
refresh is re-triggered again, but if the run is in a final state, the
refresh will only run once until the browser is focused again.