new: add shadow to the bottom of iframe in dialogs to hint scrolling
Note that the shadow fades away when the iframe is scrolled to the
bottom. However, on iframe load we do not check if the scrollHeight of
the iframe content is bigger than the iframe height, so that the bottom
shadow is always shown. This check could be done, but it might be an
overkill for a small effet, see http://stackoverflow.com/a/22991952.