new: run tags can be saved by pressing enter key
Please note that the previous behaviour of saving the tags on 'focusout'
is not honored anymore because it is not explicit. However, entering
tags but not saving them by pressing the enter key will flag the page as
'dirty' so that the user will be asked to confirm before leaving the
page.