change: pause run refresh when page is 'hidden', not on 'blur' event
The window.document.hidden boolean is set to true only when the browser
tab is not the active one, but not on focus lost. This allows to have
the browser window with the run page visible while working on another
application, but the run still updates. Only when the tab is not
visible, the refresh will be paused.

"Of course, the Page Visibility API does have its limits. Browsers do
not fire an event if the user opens a new window that covers the first.
Similarly, this API can’t tell you whether or not the browser window is
the active window. If you switch applications, your document code will
be executed as though the browser window was at the fore."

Source: http://tiffanybbrown.com/2013/02/04/page-visibility-api-tutorial/