new: show warning with explanation when clicking on a disabled action
Previously this only worked on the actions within the dropdown menu
(since they are 'a' tags) but not in the actions outside (because they
are 'button' tags). Now it works for both.

Note also that the warning shows the reason for the action being
disabled. This can be currently 2 things:
- the action is disabled because it can be performed only by super-users
- the action is disabled because the ACLs of the module don't allow the
  action to the logged-in user.