Revert "Add options to html (and error) interface"
@mebster

This reverts commit 7e5e9b576e5acad16aee872aa7f5a7e2b4a0eb35.

Commit 7e5e9 introduced changes needed for pagination. The corresponding
changes in the Server project are done in the 'feature/pagination'
branch, so that Server and UI are not in sync.

After reverting, I'll publish a branch with this commit applied again,
so that Server and UI are in sync.

@konstan

Please try to build all master branches again.