change: fixed alerts are now not dismissible
@konstan @st

You mentioned that it was annoying to see the ss:abort error alert
comming againg and againg event after having previously dismissed it.

A solution to make it less annoying is to leave the fixed alerts always
on screen and do not let the user dismiss them. After all, if we want
them to be fixed on screen it means that they are important (more
important than then floating ones).

Note that 'fixed alerts' are the ones that do take up layout space and
scroll with the content. The other alert type, called 'floating alert'
is the one that appears at the very top of the screen without changing
the layout of the page. Floating alerts can be dismissed (closed) by the
user. They also are auto-dismissed after some 5 secs (depending on the
text length) unless the mouse is hovering the alert.

I hope that makes sense. Feel free to comment.

Cheers.