new: add hidden input for 'bypass-ssh-check=true' on run dialogs, #182
This parameter is not needed for the 'Build Image' dialog, because there
is no need for SSH access when building an image. However, since this
dialog is generated from the same template than the 'Run Image' dialog,
it's still present. There is no need to explicitely removing it, because
it's ignored by the server.