change: use a custom version of bootstro.js, #238
Further customising the bootstro behaviour from the outside is becoming
too complex. Since the bootstro development looks not very active, we
fork at this point the boostro.js file.

This is a clean commit of the bootstro.js file. The only diff form the
original file is that trailing spaces are trimmed.

If we have to update bootstro.js at some point, we can diff the current
version of bootstro_custom.js with bootstro.js to see the changes that
have been applied in the meantime.