new: show total number of tour scenes, instead of page scenes, #238
In each popover (aka, step or scene) there is a label like 'x/y' on the
title showing the number progress. By default, the 'y' was the number of
steps in the page. But in multi-pages tours (e.g. Alice's tour to deploy
a wordpress) we want to show the total number of steps instead.