fix: tour was broken on Safari, #238
Reported by @mebster.

Root cause was that the function to select and open a collapsible
section (used during the tour to move to another section) was wrong:
the wanted jQuery element was selected from the jQuery array directly
with [ index - 1 ], which returns a DOM element instead of a jQuery
object. Therefore, the subsequent function call (.closest()) failed
since it .closest() is undefined on a raw DOM element.

Weirdly, this was not failing on Chrome and on Firefox.

The correct way to select only one jQuery object from a jQuery selection
is with the function .eq(zero-based-index).