new: check all european timezones abbreviations in timestamps, #308
We added more timezones abbreviations to the check, but the proper
solution is to use ISO-8601 format in the server, as noted in the issue
slipstream/SlipStreamServer#349.

At the same time the code that checks if the given timezone
abbreviations appear in the timestamp string and replaces it by the
corresponding timezone name has been enhanced to avoid to many regex
comparisions.