fix: user profile on single cloud context, #323
FYI @schaubl @anis-sixsq

The cause of the bug was the way we determined if a section is the only
section in the page, in which case we remove the collapsible handle and
we force it to remain open.

Because now sections can be split in several groups (see commit dc6e5a0e)
all group sections must be taken into account to determine if a section
is unique in the page.