fix: always send the parameter category when saving the image, #343
The input/output parameter category cannot be changed for existent
parameters, i.e. the server does not handle it properly. Therefore we
make them non-editable on the UI. However, doing so we disabled also
sending its value in the form when saving the image, leading to
inconsitent values.

Now we include a hidden input field with the value of the category, so
that it is sent as expected by the server.