change: shorten module name in runs table on dashboard, #400
@mebster @loomis

FYI

Since we are including more and more information to this table, the
columns start to take a lot of space. With this commit we shorten the
module name displayed in the runs table on the dashboard. Instead of
displaying the whole module path, i.e.:

  - Text: "first-project/newdeployment/24"
  - URL:  /module/first-project/newdeployment/24

we now only display the last module name itself and the version, with a
tooltip with the whole path as before, i.e.:

  - Text:    "newdeployment v24"
  - Tooltip: "first-project/newdeployment/24"
  - URL:     "/module/first-project/newdeployment/24"