fix: display correct number of pages when count % limit == 0, #432
@st

Previously, having e.g. a total count of 20 items with a limit of 20 per
page, displayed two pages in the pagination control. Requesting the
last page made the pagination hang forever or display a wrong empty
pages.

Now the number of pages in the pagination control is the expected one,
i.e. in the case discussed above there is only one page shown.