new: tag as :pending the vms running for less that 15 min, #553
This is to identify an unknown run-uuid in the case that the VM was
recently started, in which case SlipStream will not yet have had the
feedback about it and the run-uuid is still unkonwn.

Note that the 15min threshold is arbitrary but in sync with what the
server considers a pending vm.