fix: equally named sections can now be properly open and closed, #612
@st @loomis @schaubl

FYI I've simply added the index of the section in the section id to
make it unique.

Note that I'm not using e.g. a rand-int because the id of the open
section goes to the url when the section is open in order to let the
browser always open this section. If it was a rand-int, the section id would
not be identifiable when the page is re-loaded, e.g. when you share the
url with somebody else.